Rewrote high-level spec for parsing peptide data from a single peptide source file.  It now wants the user to select a file from the peptide_sources table, then looks for message that peptides have been uploaded.
Right now, implementation code saves the file, but we want that to be part of Peptide Source, not Peptide Parse.  So, we will take out the file save and assume the file is already saved to the directory specified by the peptide_source_path config in each environment file.

Also modified view to have select box, rather than file select.  The select box shows the rows from the peptide_sources table.